Reduce extra copy

+++ chainer/functions/connection/dilated_convolution_2d.py
@@ -69,7 +69,7 @@ class DilatedConvolution2DFunction(function.Function):
             x, kh, kw, self.sy, self.sx, self.ph, self.pw,
             cover_all=self.cover_all, dy=self.dy, dx=self.dx)
         y = numpy.tensordot(
-            self.col, W, ((1, 2, 3), (1, 2, 3))).astype(x.dtype)
+            self.col, W, ((1, 2, 3), (1, 2, 3))).astype(x.dtype, copy=False)
         if b is not None:
             y += b
         return numpy.rollaxis(y, 3, 1),
@@ -163,8 +163,8 @@ class DilatedConvolution2DFunction(function.Function):
         h, w = x.shape[2:]
 
         gW = numpy.tensordot(
-            gy, self.col, ((0, 2, 3), (0, 4, 5))).astype(W.dtype)
-        gcol = numpy.tensordot(W, gy, (0, 1)).astype(x.dtype)
+            gy, self.col, ((0, 2, 3), (0, 4, 5))).astype(W.dtype, copy=False)
+        gcol = numpy.tensordot(W, gy, (0, 1)).astype(x.dtype, copy=False)
         gcol = numpy.rollaxis(gcol, 3)
         gx = conv.col2im_cpu(gcol, self.sy, self.sx,
                              self.ph, self.pw, h, w, dy=self.dy, dx=self.dx)
